removed huge chunks in favor of many small ones
Currently this introduces a performance regression, and as such can't be merged
in this state. The rationale behind this change is that it makes code for
handling chunks a lot more straightforward, as there isn't a distinction
between master and sub chunks like there used to be. There are just chunks.

The way to go with this would be to introduce "framebuffer atlases" -
basically a way of grouping multiple framebuffers onto one large
texture. This would allow us to batch draw calls for these framebuffers.

I'm sure there is still some possibility for optimizing this without
introducing the extra complexity of framebuffer atlases, simply by
removing redundant GL API calls.